Improve sandworm_turnTowardTargetAnim: (s16) compares + shifts match target
Second range compare and the >>2 turn-scale shifts now operate on
(s16)shifted (matching target's reuse of the truncated value), dropping a
divergent region. Residual: spurious extsh on the s16 += store plus the
shifted r3/r4 placement (register/peephole cap) and the @119-vs-named
sdata2 conversion-magic reloc (#70 score-neutral).

+++ src/main/dll/dll_014C_babycloudrunner.c
@@ -671,7 +671,7 @@ void sandworm_turnTowardTargetAnim(int a, int b, u8* c, int d)
     shifted = Obj_GetYawDeltaToObject(a, b, 0) >> 3;
     *(s16*)a += shifted;
     if (d == 0) return;
-    if ((s16)shifted > -200 && shifted < 200)
+    if ((s16)shifted > -200 && (s16)shifted < 200)
     {
         if (((BabyCloudRunnerState*)c)->turnLatch != 0)
         {
@@ -695,11 +695,11 @@ void sandworm_turnTowardTargetAnim(int a, int b, u8* c, int d)
             int t;
             if ((int)(s16)shifted > 0)
             {
-                t = shifted >> 2;
+                t = (s16)shifted >> 2;
             }
             else
             {
-                t = -shifted >> 2;
+                t = -(s16)shifted >> 2;
             }
             ((ObjAnimAdvanceObjectFirstF32Fn)ObjAnim_AdvanceCurrentMove)(a, (f32)(s16)t / lbl_803E4240, timeDelta, 0);
         }
